Update cc1200.c
Think of a network as Root->Node1->Node2. In a special case where Node2 and Node1 have some cells scheduled which overlaps the autonomous cells of the Root. When Node2 transmits to Node1 in the 13th slot (which is also the root’s autonomous rx slot), the root also overhears the packet and then captures the ACK that Node1 sent back. The radio driver stores this ACK in rx_pkt (in rx_interrupt) during the same slot, but since the read() function had already been called once for the original data frame, the ACK stays stuck in the rx_pkt. In the next RX slot of the root, the rx_interrupt drops the new incoming packet because rx_pkt is still occupied by the ACK, and then read() mistakenly delivers the leftover ACK instead of the fresh frame, which results in a frame type error.
The receiver should not be receiving the ACK because its radio is supposed to be off after receiving the original packet. This issue is caused by rx_interrupt() is being called after idle() in the off() function in cc1200.c. Because rx_interrupt calls rx_rx() at the end, so the radio switches on again. After putting idle() after rx_interrupt() in off(), all the issues have been solved. The receiver is not receiving the ACK anymore, therefore no frame type errors.